Define ZyncOptionProvider interface and an implementation to Zync
A ZyncOptionsProvider interface has been defined. The intention of this
is to provide a way to build a ZyncOptions object coming from
different sources of data. The ZyncOptionsProvider implementation
in the class named CLIZyncOptionsProvider for the moment has been
placed in the component file itself but maybe will be changed
to a standalone package.