Move system-database secret creation and reconciliation in mysql component.
Also, new individual fields for system-database have been added in order
for the Operator to be able to reconcile them.
The idea with moving the Secret into mysql component is that
system should not be responsible of creating the system-database secret,
only use it. The responsibles of creating the system-database secret
should be the databases themselves.