Move system-mysql service into the 'mysql' component.
The system-mysql service has been moved into the
mysql component. This has been done because there are situations
where we don't want to deploy a mysql pod and instead use another
element, like an external database or a different database type.
By moving the system-mysql specific elements into the mysql component
we make sure that when this component is not used then this
elements are no created, and also we don't need another element
to remove them.
Also, highavailability component has been updated to create the secret
by itself because when highavailabilty compoent is used no mysql
component is used in the cluster.