Add direct client reader to Kubernetes API in apimanager reconciler
We create a Client Reader that directly queries the API server
without going to the Cache provided by the Manager's split client because
there are some resources that do not implement Watch (like ImageStreamTag)
and the Manager's Client always tries to use the Cache when reading.
This is needed at least to manage ImageStreamTag modifications during
upgrades that we will need to perform.