Move System's MySQL image into its own component
Move System's MySQL ImageStream object into the
SystemMySQLImage component. This allows us to
be able to instantiate only the MySQL ImageStream when desired.
This is needed to be able to optionally deploy it. For example,
when System's PostgreSQL is used we should not deploy the
System's MySQL ImageStream.
The MYSQL_IMAGE parameter has also been renamed
into SYSTEM_DATABASE_IMAGE