Remove operator-specific productVersion configurability
This commit removes the ability to configure the productVersion
through the APIManager CR field.
The idea is that every new release of 3scale (excluding bugfix ones)
we'll provide a new operator version. In that regard this is the same
behaviour than before. When we were to release a new 3scale version
new code was needed to be added for the new URLs so this does not change
previous behaviour in that sense.
We'll stop providing upstream nightly configurability. Instead, we'll
provide an example CR deployment with the set of nightly images(by overriding
the image fields) in case someone needs it and we won't commit to support
that at all because of the inherently unstability of those images
and the purpose of them.
We'll probably also have an upstream version of the operator but in another
branch and with stable tags and not nightly ones.
Removing this configurability that for the moment does not provide any value
allows us to not commit to maintain it. If in the future we deem necessary to
provide configurable ProductVersion through the CR we can readd it.
The upgrades of operator versions will be limited between the new version
and the older one which is the same as we have now with templates.