Add needed roles to be able to create zync-que-role in the Operator
In Kubernetes, when a new role is created the user/serviceaccount
that creates the role has to have the same permissions as the role
that it wants to create, even if he does not directly uses them.
The documentation link explaining this:
https://kubernetes.io/docs/reference/access-authn-authz/rbac/#privilege-escalation-prevention-and-bootstrapping
There's an 'escalate' permission that might allow us to prevent
this but is only compatible with Kubernetes 1.12 or newer
and at this moment we want to maintain compatibility with
at least Kubernetes 1.11.