Fix operator not being able to create service for Prometheus metrics.
The operator wasn't able to create the Service that is created
for exposing the operator Prometheus metrics. That was because
you cannot have a Service starting with a number (an error was shown).
The OPERATOR_NAME environment variable of the operator's Deployment
element seems to be the one that controls the Service name.
The "3scale-operator" OPERATOR_NAME value has been changed to
"threescale-operator" to make it compatible.
The Pod 'name' label is also used by the Service to select the pods
of the Deployment. We have also changed the pod label "name" to
"threescale-operator" because otherwise the Service did not select
any Pod.
Also an additional permission was missing to perform garbage collection
of the operator.