Ensure returned DeploymentConfigs in APIManager controller are ordered
In newer Kubernetes client versions the List method of the client
returns the elements in a random order. This was causing the
comparison of existing DeploymentConfigs return false sometimes
when setting the status.Deployments attribute of the APIManager.
We now sort them by the Name attribute after retrieving them with
the List method before comparing them.