Filter the GVKs to desired ones
We filter the existing GKVs defined in the used
runtime.Schema for the operator to a set of ones we want.
The reason for that is that kube-metrics tries to list all of
the defined Kinds in the schemas that are passed, including Kinds
that the operator doesn't use and thus the role used the
operator doesn't have them set and we don't want to set as
they are not used by the operator.