Zync BuildConfig with UBI7 base image
Changes the image of the base ImageStream used by the zync BuildConfig from centos/ruby-24-centos7 to registry.access.redhat.com/ubi7/ruby-25. This not only changes the underlying operating system to RHEL7, but also upgrades Ruby to 2.5.

This change follows https://github.com/3scale/zync/pull/328, which aligned the recommended settings for building Zync in Opesnhift with its current Dockerfile in the upstream and Ruby version upon which CI tests are run and tdownstream images are built.