Fix APIManagerRestore shared secret dependency between steps
Several restore steps always expected the shared secret
to always be available. However, one of the last steps is removing
them. That made steps fail in some situations where reconciliation
loop was reexecuted again.
This commit adds a new Status field 'APIManagerToRestoreRef' that
contains the name of the APIManager to restore, and the steps
that needed the secret now can do their tasks using this flag.