Add APIManagerBackup and APIManagerRestore CRD tests
Implement path omission mechanism in CRD tests to exclude specific fields

When used fields in the CRDs have unexported fields or are serialized
in a different format than the defined Golang object the
schema.MissingEntries method reports them as missing.
Implemented a mechanism to be able to ignore specific entries
and added the ones where we were having this situation.