Execute APIManager E2E tests with S3 configuration
A placeholder S3 configuration secret is set to allow
the deployment of APIManager success.
The S3 configuration data has been set with fake data.

The motivation of this change is to be able to run the E2E
tests on a CI system that does not support RWX(ReadWriteMany)
Persistent Volumes.