apimanager: Do not set APIManager controller-based ownerReferences on Secrets
The reason for this is that secrets are often managed by users and users
sometimes manage the secrets themselves by other controllers which also
add controller-based OwnerReferences to the secrets. In K8s only
a single controller-based OwnerReference is allowed per object, which
was preventing users to work with their own controller-based secrets.
An implication of this change is that the responsability to cleanup
the APIManager secrets deployed belongs to the user.
This changes include an upgrade procedure to remove the APIManager
controller-based ownerReferences from existing APIManager
secrets.