fix: Granular options provider logic
The `GetHighAvailabilityOptions` method that is used by all external
component reconcilers has logic that expects the secrets for all the
components being external. This means that even when a dependency is
configured as internal, as long as another dependency is configured
external this logic will try to retrieve the secret for the internal
dependency, failing the reconciliation.

Fix this by making the logic only populate the options for the
dependencies configured as external.