fix: Make Redis ConfigMap common
Currently, the Backend Redis reconciler creates a `redis-config`
ConfigMap the is used by the deployments, but the System Redis also
depends on that ConfigMap. This results in the failure to deploy the
System Redis when the Backend Redis is configured as external (which is
now possible since making this configuration fine grained).

Fix this by making the ConfigMap reconciliation common among both Redis
reconcilers, which will result in the ConfigMap being created regardless
of what combination of internal/external Redis is configured.